requirements.txt
# LLM APIs
anthropic>=0.18.0

# Core dependencies
python-dotenv>=1.0.0
gTTS>=2.5.0

# Web Dashboard
streamlit>=1.28.0

